backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
python-dotenv
redis[asyncio]
gunicorn
